[Chore] #33 CommentBar 파일 위치, enum 위치 변경
